feat: i cannot believe they can't get JSON right
or I can. I don't know. JSON is one of the simplest specs
in the world. It has literally one MUST NOT.

You MUST NOT have a byte-order-mark. It is not legal.
It has never been legal, it WILL NEVER be legal.

And yet here we are. BMS Tables sometimes spit out invalid JSON,
and you've gotta parse it. I hate BMS. Genuinely.